+++ src/util/Time.cpp
@@ -20,39 +20,39 @@ namespace pTK
         m_start = std::chrono::high_resolution_clock::now();
     }
     
-    uint32_t Time::nanoseconds() const
+    uint64 Time::nanoseconds() const
     {
         std::chrono::high_resolution_clock::time_point now = std::chrono::high_resolution_clock::now();
-        return (uint32_t)std::chrono::duration_cast<std::chrono::nanoseconds>(now - m_start).count();
+        return (uint64)std::chrono::duration_cast<std::chrono::nanoseconds>(now - m_start).count();
     }
     
-    uint32_t Time::microseconds() const
+    uint64 Time::microseconds() const
     {
         std::chrono::high_resolution_clock::time_point now = std::chrono::high_resolution_clock::now();
-        return (uint32_t)std::chrono::duration_cast<std::chrono::microseconds>(now - m_start).count();
+        return (uint64)std::chrono::duration_cast<std::chrono::microseconds>(now - m_start).count();
     }
     
-    uint32_t Time::milliseconds() const
+    uint64 Time::milliseconds() const
     {
         std::chrono::high_resolution_clock::time_point now = std::chrono::high_resolution_clock::now();
-        return (uint32_t)std::chrono::duration_cast<std::chrono::milliseconds>(now - m_start).count();
+        return (uint64)std::chrono::duration_cast<std::chrono::milliseconds>(now - m_start).count();
     }
     
-    uint32_t Time::seconds() const
+    uint64 Time::seconds() const
     {
         std::chrono::high_resolution_clock::time_point now = std::chrono::high_resolution_clock::now();
-        return (uint32_t)std::chrono::duration_cast<std::chrono::seconds>(now - m_start).count();
+        return (uint64)std::chrono::duration_cast<std::chrono::seconds>(now - m_start).count();
     }
     
-    uint32_t Time::minutes() const
+    uint32 Time::minutes() const
     {
         std::chrono::high_resolution_clock::time_point now = std::chrono::high_resolution_clock::now();
-        return (uint32_t)std::chrono::duration_cast<std::chrono::minutes>(now - m_start).count();
+        return (uint32)std::chrono::duration_cast<std::chrono::minutes>(now - m_start).count();
     }
     
-    uint32_t Time::hours() const
+    uint32 Time::hours() const
     {
         std::chrono::high_resolution_clock::time_point now = std::chrono::high_resolution_clock::now();
-        return (uint32_t)std::chrono::duration_cast<std::chrono::hours>(now - m_start).count();
+        return (uint32)std::chrono::duration_cast<std::chrono::hours>(now - m_start).count();
     }
 }
